backend/refactor: [EAI] Unify letters tables into one - models
