--- backend/models/src/permohonan/mod.rs
@@ -1,2 +1,0 @@
-pub mod generated;
-pub use generated::*;
